Expose permissions on the account serializer and reorg the code a little.
