+++ src/olympia/accounts/tests/test_serializers.py
@@ -7,10 +7,55 @@ from olympia.amo.templatetags.jinja_helpers import absolutify
 from olympia.amo.tests import (
     addon_factory, days_ago, TestCase, user_factory)
 from olympia.accounts.serializers import (
+    BaseUserSerializer,
     PublicUserProfileSerializer, LoginUserProfileSerializer,
     UserNotificationSerializer, UserProfileSerializer)
-from olympia.users.models import UserNotification
+from olympia.users.models import UserNotification, UserProfile
 from olympia.users.notifications import NOTIFICATIONS_BY_SHORT
+
+
+class TestBaseUserSerializer(TestCase):
+    serializer_class = BaseUserSerializer
+
+    def setUp(self):
+        self.request = APIRequestFactory().get('/')
+        self.user = user_factory()
+
+    def serialize(self):
+        # Manually reload the user first to clear any cached properties.
+        self.user = UserProfile.objects.get(pk=self.user.pk)
+        serializer = self.serializer_class(context={'request': self.request})
+        return serializer.to_representation(self.user)
+
+    def test_basic(self):
+        result = self.serialize()
+        assert result['id'] == self.user.pk
+        assert result['name'] == self.user.name
+        assert result['url'] is None
+
+    def test_url_for_yourself(self):
+        # should include account profile url
+        self.request.user = self.user
+        result = self.serialize()
+        assert result['url'] == absolutify(self.user.get_url_path())
+
+    def test_url_for_developers(self):
+        # should include account profile url
+        addon_factory(users=[self.user])
+        result = self.serialize()
+        assert result['url'] == absolutify(self.user.get_url_path())
+
+    def test_url_for_admins(self):
+        # should include account profile url
+        admin = user_factory()
+        self.grant_permission(admin, 'Users:Edit')
+        self.request.user = admin
+        result = self.serialize()
+        assert result['url'] == absolutify(self.user.get_url_path())
+
+    def test_username(self):
+        serialized = self.serialize()
+        assert serialized['username'] == self.user.username
 
 
 class TestPublicUserProfileSerializer(TestCase):
@@ -36,7 +81,7 @@ class TestPublicUserProfileSerializer(TestCase):
 
         self.user.update(picture_type='image/jpeg')
         serial = self.serialize()
-        assert serial['picture_url'] == self.user.picture_url
+        assert serial['picture_url'] == absolutify(self.user.picture_url)
         assert '%s.png' % self.user.id in serial['picture_url']
 
     def test_basic(self):
@@ -94,7 +139,38 @@ class TestPublicUserProfileSerializer(TestCase):
         assert data['has_anonymous_display_name'] is False
 
 
-class TestUserProfileSerializer(TestPublicUserProfileSerializer):
+class PermissionsTestMixin(object):
+    def test_permissions(self):
+        assert self.serializer(self.user).data['permissions'] == []
+
+        # Single permission
+        group = Group.objects.create(name='a', rules='Addons:Review')
+        GroupUser.objects.create(group=group, user=self.user)
+        assert self.serializer(self.user).data['permissions'] == [
+            'Addons:Review']
+
+        # Multiple permissions
+        group.update(rules='Addons:Review,Personas:Review,Addons:Edit')
+        del self.user.groups_list
+        assert self.serializer(self.user).data['permissions'] == [
+            'Addons:Edit', 'Addons:Review', 'Personas:Review']
+
+        # Change order to test sort
+        group.update(rules='Personas:Review,Addons:Review,Addons:Edit')
+        del self.user.groups_list
+        assert self.serializer(self.user).data['permissions'] == [
+            'Addons:Edit', 'Addons:Review', 'Personas:Review']
+
+        # Add a second group membership to test duplicates
+        group2 = Group.objects.create(name='b', rules='Foo:Bar,Addons:Edit')
+        GroupUser.objects.create(group=group2, user=self.user)
+        Group.objects.invalidate(*Group.objects.all())
+        assert self.serializer(self.user).data['permissions'] == [
+            'Addons:Edit', 'Addons:Review', 'Foo:Bar', 'Personas:Review']
+
+
+class TestUserProfileSerializer(TestPublicUserProfileSerializer,
+                                PermissionsTestMixin):
     serializer = UserProfileSerializer
 
     def setUp(self):
@@ -130,7 +206,7 @@ class TestUserNotificationSerializer(TestCase):
         assert data['mandatory'] == user_notification.notification.mandatory
 
 
-class TestLoginUserProfileSerializer(TestCase):
+class TestLoginUserProfileSerializer(TestCase, PermissionsTestMixin):
     serializer = LoginUserProfileSerializer
     user_kwargs = {
         'username': 'amo', 'email': 'amo@amo.amo', 'display_name': u'Ms. Amó'}
@@ -145,31 +221,3 @@ class TestLoginUserProfileSerializer(TestCase):
         assert data['name'] == self.user_kwargs['display_name']
         assert data['picture_url'] == absolutify(self.user.picture_url)
         assert data['username'] == self.user_kwargs['username']
-
-    def test_permissions(self):
-        assert self.serializer(self.user).data['permissions'] == []
-
-        # Single permission
-        group = Group.objects.create(name='a', rules='Addons:Review')
-        GroupUser.objects.create(group=group, user=self.user)
-        assert self.serializer(self.user).data['permissions'] == [
-            'Addons:Review']
-
-        # Multiple permissions
-        group.update(rules='Addons:Review,Personas:Review,Addons:Edit')
-        del self.user.groups_list
-        assert self.serializer(self.user).data['permissions'] == [
-            'Addons:Edit', 'Addons:Review', 'Personas:Review']
-
-        # Change order to test sort
-        group.update(rules='Personas:Review,Addons:Review,Addons:Edit')
-        del self.user.groups_list
-        assert self.serializer(self.user).data['permissions'] == [
-            'Addons:Edit', 'Addons:Review', 'Personas:Review']
-
-        # Add a second group membership to test duplicates
-        group2 = Group.objects.create(name='b', rules='Foo:Bar,Addons:Edit')
-        GroupUser.objects.create(group=group2, user=self.user)
-        Group.objects.invalidate(*Group.objects.all())
-        assert self.serializer(self.user).data['permissions'] == [
-            'Addons:Edit', 'Addons:Review', 'Foo:Bar', 'Personas:Review']
