--- src/olympia/users/serializers.py
@@ -1,36 +1,0 @@
-from rest_framework import serializers
-
-from olympia import amo
-from olympia.access import acl
-from olympia.amo.templatetags.jinja_helpers import absolutify
-from olympia.users.models import UserProfile
-
-
-class BaseUserSerializer(serializers.ModelSerializer):
-    url = serializers.SerializerMethodField()
-
-    class Meta:
-        model = UserProfile
-        fields = ('id', 'name', 'url', 'username')
-
-    def get_url(self, obj):
-        def is_adminish(user):
-            return (user and
-                    acl.action_allowed_user(user, amo.permissions.USERS_EDIT))
-
-        request = self.context.get('request', None)
-        current_user = getattr(request, 'user', None) if request else None
-        # Only return your own profile url, and for developers.
-        if obj == current_user or is_adminish(current_user) or obj.is_public:
-            return absolutify(obj.get_url_path())
-
-
-class AddonDeveloperSerializer(BaseUserSerializer):
-    picture_url = serializers.SerializerMethodField()
-
-    class Meta(BaseUserSerializer.Meta):
-        fields = ('id', 'name', 'url', 'username', 'picture_url')
-        read_only_fields = fields
-
-    def get_picture_url(self, obj):
-        return absolutify(obj.picture_url)
